+++ backend/main-service/src/main/java/kkakka/mainservice/product/application/ProductService.java
@@ -7,10 +7,9 @@ import kkakka.mainservice.product.application.dto.CategoryDto;
 import kkakka.mainservice.product.application.dto.NutritionDto;
 import kkakka.mainservice.product.application.dto.ProductDetailDto;
 import kkakka.mainservice.product.application.dto.ProductDto;
-import kkakka.mainservice.product.application.recommend.ProductRecommender;
 import kkakka.mainservice.product.application.recommend.RecommendStrategy;
-import kkakka.mainservice.product.application.recommend.RecommendStrategyFactory;
 import kkakka.mainservice.product.application.recommend.RecommenderFactory;
+import kkakka.mainservice.product.application.recommend.strategy.ProductRecommender;
 import kkakka.mainservice.product.domain.Product;
 import kkakka.mainservice.product.domain.SearchWords;
 import kkakka.mainservice.product.domain.repository.ProductRepository;
