+++ backend/main-service/src/main/java/kkakka/mainservice/product/application/recommend/RecommendProductIds.java
@@ -1,16 +1,20 @@
-package kkakka.mainservice.product.application.recommend.dto;
+package kkakka.mainservice.product.application.recommend;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class RecommendProductDto {
+public class RecommendProductIds {
 
+    @Id
+    @JsonProperty("id")
+    private String pivotId;
     @JsonProperty("product_ids")
     private List<Long> productIds;
 }
