+++ backend/main-service/src/main/java/kkakka/mainservice/product/application/recommend/RecommendStrategyFactory.java
@@ -4,6 +4,10 @@ import java.util.HashMap;
 import java.util.Map;
 import javax.annotation.PostConstruct;
 import kkakka.mainservice.member.auth.ui.Authority;
+import kkakka.mainservice.product.application.recommend.strategy.AnonymousRecommendStrategy;
+import kkakka.mainservice.product.application.recommend.strategy.MemberRecommendStrategy;
+import kkakka.mainservice.product.application.recommend.strategy.ProductRecommendStrategy;
+import kkakka.mainservice.product.application.recommend.strategy.ProductRecommender;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
