+++ backend/main-service/src/main/java/kkakka/mainservice/product/application/recommend/RecommenderFactory.java
@@ -1,6 +1,7 @@
 package kkakka.mainservice.product.application.recommend;
 
 import kkakka.mainservice.member.auth.ui.Authority;
+import kkakka.mainservice.product.application.recommend.strategy.ProductRecommender;
 
 public interface RecommenderFactory {
