+++ backend/main-service/src/main/java/kkakka/mainservice/product/application/recommend/strategy/MemberRecommendStrategy.java
@@ -1,10 +1,12 @@
-package kkakka.mainservice.product.application.recommend;
+package kkakka.mainservice.product.application.recommend.strategy;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Random;
 import java.util.stream.Collectors;
@@ -15,7 +17,7 @@ import kkakka.mainservice.member.member.domain.repository.MemberRepository;
 import kkakka.mainservice.order.domain.Order;
 import kkakka.mainservice.order.domain.ProductOrder;
 import kkakka.mainservice.order.domain.repository.OrderRepository;
-import kkakka.mainservice.product.application.recommend.dto.RecommendProductDto;
+import kkakka.mainservice.product.application.recommend.RecommendProductIds;
 import kkakka.mainservice.product.domain.Product;
 import kkakka.mainservice.product.domain.repository.ProductRepository;
 import kkakka.mainservice.review.domain.Review;
@@ -84,16 +86,17 @@ public class MemberRecommendStrategy implements ProductRecommender {
         final int randomIdx = selectRandomPivotIndex(orderedProducts.size());
 
         final Product pivotProduct = orderedProducts.get(randomIdx);
-        final RecommendProductDto recommendProductDto = requestRecommendation(pivotProduct);
+        final RecommendProductIds recommendProductIds = requestRecommendation(pivotProduct);
 
-        return findRecommendedProducts(pageable, recommendProductDto);
+        return findRecommendedProducts(pageable, recommendProductIds);
     }
 
-    private Page<Product> recommendWithHighRatedReviewRandom(Pageable pageable, List<Review> topRatedReview) {
+    private Page<Product> recommendWithHighRatedReviewRandom(Pageable pageable,
+            List<Review> topRatedReview) {
         final Product pivotProduct = reviewedPivotProduct(topRatedReview);
-        final RecommendProductDto recommendProductDto = requestRecommendation(pivotProduct);
+        final RecommendProductIds recommendProductIds = requestRecommendation(pivotProduct);
 
-        return findRecommendedProducts(pageable, recommendProductDto);
+        return findRecommendedProducts(pageable, recommendProductIds);
     }
 
     private Product reviewedPivotProduct(List<Review> topRatedReview) {
@@ -112,7 +115,7 @@ public class MemberRecommendStrategy implements ProductRecommender {
         ).getContent();
     }
 
-    private RecommendProductDto requestRecommendation(Product pivotProduct) {
+    private RecommendProductIds requestRecommendation(Product pivotProduct) {
         final ResponseEntity<String> response = restTemplate.exchange(
                 RECOMMENDATION_SERVER_URL + pivotProduct.getId(),
                 HttpMethod.GET,
@@ -120,7 +123,7 @@ public class MemberRecommendStrategy implements ProductRecommender {
                 String.class
         );
 
-        return convertResponseBody(response);
+        return convertResponseBody(pivotProduct.getId(), response);
     }
 
     private int selectRandomPivotIndex(int itemSize) {
@@ -133,9 +136,9 @@ public class MemberRecommendStrategy implements ProductRecommender {
 
     private PageImpl<Product> findRecommendedProducts(
             Pageable pageable,
-            RecommendProductDto recommendProductDto
+            RecommendProductIds recommendProductIds
     ) {
-        final List<Long> productIds = recommendProductDto.getProductIds();
+        final List<Long> productIds = recommendProductIds.getProductIds();
         final List<Long> pagedProductIds = productIds.subList(
                 (int) pageable.getOffset(),
                 (int) pageable.getOffset() + maximumPageSize(pageable, productIds)
@@ -167,11 +170,14 @@ public class MemberRecommendStrategy implements ProductRecommender {
         return Math.min(pageable.getPageSize(), (int) (productIds.size() - pageable.getOffset()));
     }
 
-    private RecommendProductDto convertResponseBody(ResponseEntity<String> response) {
+    private RecommendProductIds convertResponseBody(Long pivotId, ResponseEntity<String> response) {
         try {
+            if (Objects.isNull(response.getBody())) {
+                return new RecommendProductIds(pivotId.toString(), Collections.emptyList());
+            }
             return objectMapper.readValue(
                     response.getBody(),
-                    new TypeReference<RecommendProductDto>() {
+                    new TypeReference<RecommendProductIds>() {
                     }
             );
         } catch (JsonProcessingException e) {
