+++ backend/main-service/src/main/java/kkakka/mainservice/product/application/recommend/strategy/ProductRecommendStrategy.java
@@ -1,4 +1,4 @@
-package kkakka.mainservice.product.application.recommend;
+package kkakka.mainservice.product.application.recommend.strategy;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -10,7 +10,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import kkakka.mainservice.common.exception.InvalidRecommendResponseException;
-import kkakka.mainservice.product.application.recommend.dto.RecommendProductDto;
+import kkakka.mainservice.product.application.recommend.RecommendProductIds;
 import kkakka.mainservice.product.domain.Product;
 import kkakka.mainservice.product.domain.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
@@ -52,12 +52,12 @@ public class ProductRecommendStrategy implements ProductRecommender {
         }
 
         final Product pivotProduct = product.get();
-        final RecommendProductDto recommendProductDto = requestRecommendation(pivotProduct);
+        final RecommendProductIds recommendProductIds = requestRecommendation(pivotProduct);
 
-        return findRecommendedProducts(pageable, recommendProductDto);
+        return findRecommendedProducts(pageable, recommendProductIds);
     }
 
-    private RecommendProductDto requestRecommendation(Product pivotProduct) {
+    private RecommendProductIds requestRecommendation(Product pivotProduct) {
         final ResponseEntity<String> response = restTemplate.exchange(
                 RECOMMENDATION_SERVER_URL + pivotProduct.getId(),
                 HttpMethod.GET,
@@ -65,17 +65,17 @@ public class ProductRecommendStrategy implements ProductRecommender {
                 String.class
         );
 
-        return convertResponseBody(response);
+        return convertResponseBody(pivotProduct.getId(), response);
     }
 
-    private RecommendProductDto convertResponseBody(ResponseEntity<String> response) {
+    private RecommendProductIds convertResponseBody(Long pivotId, ResponseEntity<String> response) {
         try {
             if (Objects.isNull(response.getBody())) {
-                return new RecommendProductDto(Collections.emptyList());
+                return new RecommendProductIds(pivotId.toString(), Collections.emptyList());
             }
             return objectMapper.readValue(
                     response.getBody(),
-                    new TypeReference<RecommendProductDto>() {
+                    new TypeReference<RecommendProductIds>() {
                     }
             );
         } catch (JsonProcessingException e) {
@@ -85,9 +85,9 @@ public class ProductRecommendStrategy implements ProductRecommender {
 
     private PageImpl<Product> findRecommendedProducts(
             Pageable pageable,
-            RecommendProductDto recommendProductDto
+            RecommendProductIds recommendProductIds
     ) {
-        final List<Long> productIds = recommendProductDto.getProductIds();
+        final List<Long> productIds = recommendProductIds.getProductIds();
         final List<Long> pagedProductIds = productIds.subList(
                 (int) pageable.getOffset(),
                 (int) pageable.getOffset() + maximumPageSize(pageable, productIds)
